+++ src/main/java/org/opensearch/knn/index/util/KNNClusterUtil.java
@@ -6,9 +6,9 @@
 package org.opensearch.knn.index.util;
 
 import lombok.AccessLevel;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
+import lombok.Setter;
 import lombok.extern.log4j.Log4j2;
 import org.opensearch.Version;
 import org.opensearch.action.IndicesRequest;
@@ -16,13 +16,11 @@ import org.opensearch.cluster.metadata.IndexMetadata;
 import org.opensearch.cluster.metadata.IndexNameExpressionResolver;
 import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.core.index.Index;
+import org.opensearch.search.pipeline.SearchPipelineService;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
-
-import static org.opensearch.search.pipeline.SearchPipelineService.ENABLED_SYSTEM_GENERATED_FACTORIES_SETTING;
 
 /**
  * Class abstracts information related to underlying OpenSearch cluster
@@ -34,8 +32,8 @@ public class KNNClusterUtil {
     private ClusterService clusterService;
     private static KNNClusterUtil instance;
     private IndexNameExpressionResolver indexNameExpressionResolver;
-    @Getter
-    private List<String> enabledSystemGeneratedFactories = Collections.emptyList();
+    @Setter
+    private SearchPipelineService searchPipelineService;
 
     /**
      * Return instance of the cluster context, must be initialized first for proper usage
@@ -56,12 +54,6 @@ public class KNNClusterUtil {
     public void initialize(final ClusterService clusterService, final IndexNameExpressionResolver indexNameExpressionResolver) {
         this.clusterService = clusterService;
         this.indexNameExpressionResolver = indexNameExpressionResolver;
-        this.enabledSystemGeneratedFactories = clusterService.getClusterSettings().get(ENABLED_SYSTEM_GENERATED_FACTORIES_SETTING);
-        clusterService.getClusterSettings()
-            .addSettingsUpdateConsumer(
-                ENABLED_SYSTEM_GENERATED_FACTORIES_SETTING,
-                factories -> enabledSystemGeneratedFactories = factories
-            );
     }
 
     /**
@@ -90,5 +82,17 @@ public class KNNClusterUtil {
         return Arrays.stream(concreteIndices)
             .map(concreteIndex -> clusterService.state().metadata().index(concreteIndex))
             .collect(Collectors.toList());
+    }
+
+    /**
+     * Check if the system generated search processor factory is enabled or not
+     * @param factoryName name of the factory
+     * @return If the factory is enabled or not
+     */
+    public boolean isSystemGeneratedSearchFactoryEnabled(String factoryName) {
+        if (searchPipelineService == null) {
+            throw new IllegalStateException("search pipeline service is not initialized in the KNN cluster util.");
+        }
+        return searchPipelineService.isSystemGeneratedFactoryEnabled(factoryName);
     }
 }